backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.10.4
httpx==0.27.0
beautifulsoup4==4.12.3
langchain==0.1.12
langchain-google-genai==1.0.3
tenacity==8.2.3
structlog==24.1.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
slowapi==0.1.9

